requirements.txt
fastapi
mangum
openpyxl
boto3
python-dotenv
requests
uvicorn

